fix/skill_id_for_priority
mycroft uses the skill name, this is not deterministic!

msm decides what the name is based on the meta info from selene, if missing it uses folder name (skill_id)

for backwards compat name is still support but skill_id is recommended!
the name can be changed at any time and mess up the .conf, if the skill_id changes lots of other stuff will break so you are assured to notice

authored-by: jarbasai <jarbasai@mailfence.com>